requirements.txt
# Core Application Framework
PySide6
PySide6-Addons
PySide6-Essentials

# OCR and Document Processing
python-doctr
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
PyMuPDF
Pillow
piexif
python-docx
python-pptx
openpyxl

# Settings and Logic
pydantic
PyYAML
python-dotenv
Jinja2

# macOS Specific (for foreground activation)
pyobjc-core; sys_platform == 'darwin'
pyobjc-framework-Cocoa; sys_platform == 'darwin'

# Development & Testing
pytest
pytest-mock
pytest-qt
pytest-asyncio
pytest-cov

typing_extensions
urllib3
ollama
openai
mistralai
